Major commit, fixed many random todos, but mainly redid the following architectures:
Redid a lot of the command executer including event processing and movement processing.
Redid the way directions are handled which included a huge revamp to doorways, movement in general, viewing, npc movement etc.

These changes refactored a lot of startup code and room generation/link generation code as well as reduced and improved the code for many adjacent areas.


git-svn-id: https://subversion.assembla.com/svn/dmud.8th_Circle/trunk@83 8e71f4b0-c50f-4f4d-bad6-908c98c512d7